benchmarker now uses ZedGraph - this makes graphing linux mono compatible (and graphs zoomeable)
svn path=/trunk/other/; revision=10131

+++ Benchmarker/BatchRunResult.cs
@@ -144,7 +144,7 @@ namespace Benchmarker
 
         public class ValueEntry
         {
-            public int GameFrame;
+            public double GameFrame;
             public string Key;
             public double Value;
         }
